#1288: Refactor isStateless check on init. An appdef is now stateless if:  1. The Stateless extension property is true  2. -OR- No Map Definitions references were found in the app def
